--- src/main/java/org/example/becoco/auth/controller/AuthController.java
@@ -1,28 +1,0 @@
-package org.example.becoco.auth.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.example.becoco.auth.Service.AuthService;
-import org.example.becoco.auth.dto.LoginRequest;
-import org.example.becoco.auth.dto.TokenResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/auth")
-@RequiredArgsConstructor
-public class AuthController {
-    private final AuthService authService;
-
-    @PostMapping("/login")
-    public ResponseEntity<TokenResponse> login(@RequestBody LoginRequest loginRequest) {
-        TokenResponse tokenResponse = authService.login(loginRequest);
-        return ResponseEntity.ok(tokenResponse);
-    }
-
-    @PostMapping("/signup")
-    public void signup(@RequestBody LoginRequest loginRequest) {
-        authService.signup(loginRequest);
-    }
-}
