--- src/main/java/org/example/becoco/auth/dto/LoginRequest.java
@@ -1,12 +1,0 @@
-package org.example.becoco.auth.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class LoginRequest {
-    private String userName;
-    private String password;
-    private String role;
-}
