--- src/main/java/org/example/becoco/auth/dto/TokenResponse.java
@@ -1,13 +1,0 @@
-package org.example.becoco.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class TokenResponse {
-    private String accessToken;
-    private Long accessTokenTime;
-}
