--- src/main/java/org/example/becoco/auth/entity/Auth.java
@@ -1,29 +1,0 @@
-package org.example.becoco.auth.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-
-@Entity
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class Auth {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    private String userName;
-    private String password;
-    private String role;
-
-    public void getUserName(String userName) {
-        this.userName = userName;
-    }
-}
