--- src/main/java/org/example/becoco/auth/repository/AuthRepository.java
@@ -1,12 +1,0 @@
-package org.example.becoco.auth.repository;
-
-import org.example.becoco.auth.entity.Auth;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.security.core.userdetails.User;
-
-import java.util.Optional;
-
-public interface AuthRepository extends JpaRepository<Auth, Long> {
-    Optional<Auth> findByUserName(String userName);
-}
