--- src/main/java/org/example/becoco/auth/util/config/FilterConfig.java
@@ -1,24 +1,0 @@
-package org.example.becoco.auth.util.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
-import org.example.becoco.auth.util.jwt.JwtProvider;
-import org.example.becoco.auth.util.filter.JwtTokenFilter;
-import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.DefaultSecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class FilterConfig extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
-
-    private final JwtProvider jwtProvider;
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void configure(HttpSecurity builder) throws Exception {
-        builder.addFilterBefore(new JwtTokenFilter(jwtProvider), UsernamePasswordAuthenticationFilter.class);
-    }
-}
