--- src/main/java/org/example/becoco/auth/util/error/entity/ErrorResponseEntity.java
@@ -1,29 +1,0 @@
-package org.example.becoco.auth.util.error.entity;
-
-import lombok.Builder;
-import lombok.Getter;
-import org.example.becoco.auth.util.error.ErrorCode;
-import org.springframework.http.ResponseEntity;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-public class ErrorResponseEntity {
-    private int status;
-    private String code;
-    private String message;
-    private LocalDateTime createLocalDateTime;
-
-    public static ResponseEntity<ErrorResponseEntity> responseEntity(ErrorCode errorCode) {
-        return ResponseEntity
-                .status(errorCode.getHttpStatus())
-                .body(ErrorResponseEntity.builder()
-                        .status(errorCode.getHttpStatus())
-                        .code(errorCode.name())
-                        .message(errorCode.getMessage())
-                        .createLocalDateTime(LocalDateTime.now())
-                        .build()
-                );
-    }
-}
