--- src/main/java/org/example/becoco/auth/util/error/exception/CustomException.java
@@ -1,11 +1,0 @@
-package org.example.becoco.auth.util.error.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.example.becoco.auth.util.error.ErrorCode;
-
-@Getter
-@AllArgsConstructor
-public class CustomException extends RuntimeException{
-    private ErrorCode errorCode;
-}
