+++ src/main/java/org/example/becoco/domain/auth/dto/LoginRequest.java
@@ -1,0 +1,12 @@
+package org.example.becoco.domain.auth.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class LoginRequest {
+    private String userName;
+    private String password;
+    private String role;
+}
