+++ src/main/java/org/example/becoco/domain/auth/dto/TokenResponse.java
@@ -1,0 +1,13 @@
+package org.example.becoco.domain.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class TokenResponse {
+    private String accessToken;
+    private Long accessTokenTime;
+}
