+++ src/main/java/org/example/becoco/domain/auth/exception/ExpiredTokenException.java
@@ -1,0 +1,13 @@
+package org.example.becoco.domain.auth.exception;
+
+import org.example.becoco.global.error.ErrorCode;
+import org.example.becoco.global.error.exception.CustomException;
+
+public class ExpiredTokenException extends CustomException {
+
+    public static final CustomException EXCEPTION = new ExpiredTokenException();
+
+    private ExpiredTokenException(){
+        super(ErrorCode.JWT_EXPIRED);
+    }
+}
