+++ src/main/java/org/example/becoco/domain/auth/exception/UserMissMatchException.java
@@ -1,0 +1,10 @@
+package org.example.becoco.domain.auth.exception;
+
+import org.example.becoco.global.error.ErrorCode;
+import org.example.becoco.global.error.exception.CustomException;
+
+public class UserMissMatchException extends CustomException {
+    public static final CustomException EXCEPTION = new UserMissMatchException();
+
+    public UserMissMatchException() {super (ErrorCode.USER_MISS_MATCH);}
+}
