+++ src/main/java/org/example/becoco/domain/auth/repository/AuthRepository.java
@@ -1,0 +1,10 @@
+package org.example.becoco.domain.auth.repository;
+
+import org.example.becoco.domain.auth.entity.Auth;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface AuthRepository extends JpaRepository<Auth, Long> {
+    Optional<Auth> findByUserName(String userName);
+}
