+++ src/main/java/org/example/becoco/global/error/entity/ErrorResponseEntity.java
@@ -1,0 +1,29 @@
+package org.example.becoco.global.error.entity;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.example.becoco.global.error.ErrorCode;
+import org.springframework.http.ResponseEntity;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Builder
+public class ErrorResponseEntity {
+    private int status;
+    private String code;
+    private String message;
+    private LocalDateTime createLocalDateTime;
+
+    public static ResponseEntity<ErrorResponseEntity> responseEntity(ErrorCode errorCode) {
+        return ResponseEntity
+                .status(errorCode.getHttpStatus())
+                .body(ErrorResponseEntity.builder()
+                        .status(errorCode.getHttpStatus())
+                        .code(errorCode.name())
+                        .message(errorCode.getMessage())
+                        .createLocalDateTime(LocalDateTime.now())
+                        .build()
+                );
+    }
+}
