+++ src/main/java/org/example/becoco/global/error/handler/CustomExceptionHandler.java
@@ -1,0 +1,16 @@
+package org.example.becoco.global.error.handler;
+
+import org.example.becoco.global.error.entity.ErrorResponseEntity;
+import org.example.becoco.global.error.exception.CustomException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class CustomExceptionHandler {
+    @ExceptionHandler(CustomException.class)
+    protected ResponseEntity<ErrorResponseEntity> handlerCustomException(CustomException customException){
+        return ErrorResponseEntity.responseEntity(customException.getErrorCode());
+    }
+
+}
