+++ src/main/java/org/example/becoco/global/security/config/FilterConfig.java
@@ -1,0 +1,24 @@
+package org.example.becoco.global.security.config;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import org.example.becoco.global.security.jwt.JwtProvider;
+import org.example.becoco.global.filter.JwtTokenFilter;
+import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.web.DefaultSecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class FilterConfig extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
+
+    private final JwtProvider jwtProvider;
+    private final ObjectMapper objectMapper;
+
+    @Override
+    public void configure(HttpSecurity builder) throws Exception {
+        builder.addFilterBefore(new JwtTokenFilter(jwtProvider), UsernamePasswordAuthenticationFilter.class);
+    }
+}
