+++ src/main/java/org/example/becoco/global/security/jwt/JwtProperties.java
@@ -1,0 +1,18 @@
+package org.example.becoco.global.security.jwt;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Getter
+@Setter
+@Configuration
+@ConfigurationProperties("jwt")
+public class JwtProperties {
+    private Long accessTokenTime;
+    private String prefix;
+    private String header;
+    private String secretKey;
+    private Long exp;
+}
